Moved more code to source code files and start grouping code by object type
