+++ module/bafs/core.c
@@ -8,9 +8,8 @@
 
 #include <linux/bafs.h>
 
-#include <linux/bafs/group.h>
+#include <linux/bafs/types.h>
 #include <linux/bafs/util.h>
-#include <linux/bafs/release.h>
 
 
 MODULE_LICENSE("GPL");
@@ -23,118 +22,26 @@ MODULE_VERSION("0.1");
 dev_t bafs_major = {0};
 
 static struct class* bafs_core_class  = NULL;
-struct class*        bafs_ctrl_class  = NULL;
-struct class*        bafs_group_class = NULL;
 
 static struct cdev bafs_core_cdev;
 struct device*     bafs_core_device = NULL;
 
-DEFINE_IDA(bafs_minor_ida);
-DEFINE_IDA(bafs_ctrl_ida);
-DEFINE_IDA(bafs_group_ida);
 
-
-
-static int bafs_ctrl_pci_probe(struct pci_dev* pdev, const struct pci_device_id* id) {
-
+static int bafs_ctrl_pci_probe(struct pci_dev* pdev, const struct pci_device_id* id)
+{
     int ret = 0;
-
-    struct bafs_ctrl* ctrl;
+    struct bafs_ctrl * ctrl;
 
     BAFS_CTRL_DEBUG("Started PCI probe for PCI device: %02x:%02x.%1x\n", pdev->bus->number, PCI_SLOT(pdev->devfn), PCI_FUNC(pdev->devfn));
 
-    ctrl    = kzalloc(sizeof(*ctrl), GFP_KERNEL);
-    if (!ctrl) {
-        ret = -ENOMEM;
-        goto out;
+    ret = bafs_ctrl_alloc(&ctrl, pdev, bafs_major, bafs_core_device);
+    if(ret == 0) {
+        BAFS_CTRL_INFO("Created controller device %s for PCI device: %02x:%02x.%1x\n",
+                       dev_name(ctrl->device), pdev->bus->number, PCI_SLOT(pdev->devfn),
+                       PCI_FUNC(pdev->devfn));
     }
 
-    spin_lock_init(&ctrl->lock);
-    INIT_LIST_HEAD(&ctrl->group_list);
-
-
-    ctrl->pdev = pdev;
-    ctrl->dev  = get_device(&ctrl->pdev->dev);
-
-
-    pci_set_master(pdev);
-    pci_set_drvdata(pdev, ctrl);
-    pci_free_irq_vectors(pdev);
-    pci_disable_msi(pdev);
-    pci_disable_msix(pdev);
-    dma_set_mask_and_coherent(&pdev->dev, DMA_BIT_MASK(64));
-
-    ret = pci_request_region(pdev, 0, BAFS_CTRL_CLASS_NAME);
-    if (ret < 0) {
-        goto out_clear_pci_drvdata;
-    }
-
-    ret = pci_enable_device(pdev);
-    if (ret < 0) {
-        goto out_release_pci_region;
-    }
-
-
-
-    ret = ida_simple_get(&bafs_minor_ida, 1, BAFS_MINORS, GFP_KERNEL);
-    if (ret < 0) {
-        BAFS_CTRL_ERR("Failed to get minor instance id \t err = %d\n", ret);
-        goto out_disable_pci_device;
-    }
-    ctrl->minor = ret;
-
-    ret = ida_simple_get(&bafs_ctrl_ida, 0, 0, GFP_KERNEL);
-    if (ret < 0) {
-        BAFS_CTRL_ERR("Failed to get ctrl instance id \t err = %d\n", ret);
-        goto out_release_minor_instance;
-    }
-    ctrl->ctrl_id = ret;
-
-    bafs_ctrl_init(ctrl);
-    ret              = cdev_add(&ctrl->cdev, MKDEV(MAJOR(bafs_major), ctrl->minor), 1);
-    if (ret < 0) {
-        BAFS_CTRL_ERR("Failed to init ctrl cdev \t err = %d\n", ret);
-        goto out_release_ctrl_instance;
-    }
-    ctrl->core_dev = get_device(bafs_core_device);
-    ctrl->device   = device_create(bafs_ctrl_class, bafs_core_device, MKDEV(MAJOR(bafs_major), ctrl->minor), ctrl, BAFS_CTRL_DEVICE_NAME, ctrl->ctrl_id);
-    if (IS_ERR(ctrl->device)) {
-        ret        = PTR_ERR(ctrl->device);
-        BAFS_CORE_ERR("Failed to create ctrl device \t err = %d\n", ret);
-        goto out_delete_device_cdev;
-    }
-
-    kref_init(&ctrl->ref);
-
-    BAFS_CTRL_INFO("Created controller device %s for PCI device: %02x:%02x.%1x\n",
-                   dev_name(ctrl->device), pdev->bus->number, PCI_SLOT(pdev->devfn), PCI_FUNC(pdev->devfn));
-
-
-
-    //BAFS_CTRL_DEBUG("Finished PCI probe for PCI device: %02x:%02x.%1x\n", pdev->bus->number, PCI_SLOT(pdev->devfn), PCI_FUNC(pdev->devfn));
     return ret;
-
-out_delete_device_cdev:
-    put_device(ctrl->core_dev);
-    cdev_del(&ctrl->cdev);
-out_release_ctrl_instance:
-    ida_simple_remove(&bafs_ctrl_ida, ctrl->ctrl_id);
-out_release_minor_instance:
-    ida_simple_remove(&bafs_minor_ida, ctrl->minor);
-out_disable_pci_device:
-    pci_disable_device(pdev);
-out_release_pci_region:
-    pci_release_region(pdev, 0);
-out_clear_pci_drvdata:
-    pci_set_drvdata(pdev, NULL);
-//out_clear_pci_master:
-    pci_clear_master(pdev);
-
-//out_delete_ctrl:
-    kfree(ctrl);
-out:
-    return ret;
-
 }
 
 static void bafs_ctrl_pci_remove(struct pci_dev* pdev) {
@@ -161,17 +68,14 @@ static struct pci_driver bafs_ctrl_pci_driver = {
 
 };
 
-long bafs_core_create_group(void __user* user_params) {
+long bafs_core_create_group(void __user* user_params)
+{
+    long ret;
+    int i;
 
-    long ret = 0;
-    long i   = 0;
-    long j   = 0;
-
-    struct bafs_group*                       group;
+    struct bafs_group *                      group;
     struct BAFS_CORE_IOC_CREATE_GROUP_PARAMS params_;
     struct BAFS_CORE_IOC_CREATE_GROUP_PARAMS params;
-    struct device**                          ctrl_devices;
-    const char*                              device_class_name;
 
     if (copy_from_user(&params_, user_params, sizeof(params_))) {
         ret = -EFAULT;
@@ -184,154 +88,48 @@ long bafs_core_create_group(void __user* user_params) {
     if (!params.ctrls) {
         ret        = -ENOMEM;
         BAFS_CORE_ERR("Failed to allocate memory for bafs_group ctrl names\n");
-        goto out_free_local_ctrls;
+        goto out_free_params;
     }
-
-
 
     for (i = 0; i < params.n_ctrls; i++) {
         ret = strncpy_from_user(params.ctrls[i], params_.ctrls[i], MAX_NAME_LEN);
         if (ret < 0) {
             BAFS_CORE_ERR("Failed to copy ctrl name\n");
-            goto out_free_local_ctrls;
+            goto out_free_params;
         }
         else if (ret == MAX_NAME_LEN) {
             ret = -EINVAL;
             BAFS_CORE_ERR("Failed to copy ctrl name, too long\n");
-            goto out_free_local_ctrls;
+            goto out_free_params;
 
         }
     }
 
-    ctrl_devices = kzalloc(params.n_ctrls*sizeof(*ctrl_devices), GFP_KERNEL);
-    if (!params.ctrls) {
-        ret      = -ENOMEM;
-        BAFS_CORE_ERR("Failed to allocate memory for ctrl devices\n");
-        goto out_free_local_ctrls;
+    ret = bafs_group_alloc(&group, bafs_major, bafs_core_device, params.n_ctrls, params.ctrls);
+    if(ret < 0) {
+        goto out_free_params;
     }
 
-    get_device(bafs_core_device);
-    for (i = 0; i < params.n_ctrls; i++) {
-        ctrl_devices[i] = device_find_child_by_name(bafs_core_device, params.ctrls[i]);
-        if (!ctrl_devices) {
-            BAFS_CORE_ERR("Failed to find ctrl device: %s\n", params.ctrls[i]);
-            goto out_clean_ctrl_devices;
-        }
-        device_class_name = ctrl_devices[i]->class ? ctrl_devices[i]->class->name : "";
-        ret               = strncmp(device_class_name, BAFS_CTRL_CLASS_NAME, strlen(BAFS_CTRL_CLASS_NAME));
-        if (ret           != 0) {
-            BAFS_CORE_ERR("Failed to find ctrl device: %s\n", params.ctrls[i]);
-            goto out_clean_ctrl_devices;
-        }
 
-    }
-
-    group   = kzalloc(sizeof(*group), GFP_KERNEL);
-    if (!group) {
-        ret = -ENOMEM;
-        BAFS_CORE_ERR("Failed to allocate memory for bafs_group\n");
-        goto out_clean_ctrl_devices;
-    }
-    spin_lock_init(&group->lock);
-
-    group->ctrls = kzalloc(params.n_ctrls*sizeof(*(group->ctrls)), GFP_KERNEL);
-    if (!group->ctrls) {
-        ret      = -ENOMEM;
-        BAFS_CORE_ERR("Failed to allocate memory for bafs_group\n");
-        goto out_delete_group;
-    }
-
-    for (j = 0; j < params.n_ctrls; j++) {
-
-        group->ctrls[j] = (struct bafs_ctrl*) dev_get_drvdata(ctrl_devices[j]);
-        if (!group->ctrls[j]) {
-            ret = -EFAULT;
-            BAFS_CORE_ERR("Failed to find ctrl device: %s\n", params.ctrls[j]);
-            goto out_delete_group_ctrls;
-        }
-        bafs_get_ctrl(group->ctrls[j]);
-    }
-
-    ret = ida_simple_get(&bafs_minor_ida, 1, BAFS_MINORS, GFP_KERNEL);
-    if (ret < 0) {
-        BAFS_CORE_ERR("Failed to get minor instance id \t err = %ld\n", ret);
-        goto out_delete_group_ctrls;
-    }
-    group->minor = ret;
-
-    ret = ida_simple_get(&bafs_group_ida, 0, 0, GFP_KERNEL);
-    if (ret < 0) {
-        BAFS_CORE_ERR("Failed to get group instance id \t err = %ld\n", ret);
-        goto out_release_minor_instance;
-    }
-    group->group_id = ret;
-    group->n_ctrls  = params.n_ctrls;
-
-    //cdev_init(&group->cdev, &bafs_group_fops);
-    cdev_init(&group->cdev, NULL);
-    group->cdev.owner = THIS_MODULE;
-    ret               = cdev_add(&group->cdev, MKDEV(MAJOR(bafs_major), group->minor), 1);
-    if (ret < 0) {
-        BAFS_CORE_ERR("Failed to init group cdev \t err = %ld\n", ret);
-        goto out_release_group_instance;
-    }
-
-    group->core_dev = bafs_core_device;
-
-
-    group->device = device_create(bafs_group_class, bafs_core_device, MKDEV(MAJOR(bafs_major), group->minor), group, BAFS_GROUP_DEVICE_NAME, group->group_id);
-    if (IS_ERR(group->device)) {
-        ret       = PTR_ERR(group->device);
-        BAFS_CORE_ERR("Failed to create group device \t err = %ld\n", ret);
-        goto out_delete_device_cdev;
-    }
-    kref_init(&group->ref);
-    BAFS_CORE_INFO("Created group device %s\n", dev_name(group->device));
-
-    ret      = snprintf(params.group_name, MAX_NAME_LEN, "/dev%s", dev_name(group->device));
+    ret = snprintf(params.group_name, MAX_NAME_LEN, "/dev%s", dev_name(group->device));
     if (ret != 0) {
         ret  = -EINVAL;
         BAFS_CORE_ERR("Failed to copy group name\n");
-        goto out_destroy_device;
+        goto out_free_group;
     }
 
     if (copy_to_user(user_params, &params, sizeof(params))) {
         ret = -EFAULT;
         BAFS_CORE_ERR("Failed to copy params to user\n");
-        goto out_destroy_device;
+        goto out_free_group;
     }
 
-    for (i = 0; i < params.n_ctrls; i++) {
-        put_device(ctrl_devices[i]);
-    }
+    kfree(params.ctrls);
 
-    kfree(ctrl_devices);
-
-    ret = 0;
-    return ret;
-
-out_destroy_device:
-    device_destroy(bafs_group_class, MKDEV(MAJOR(bafs_major), group->minor));
-out_delete_device_cdev:
-    cdev_del(&group->cdev);
-out_release_group_instance:
-    ida_simple_remove(&bafs_group_ida, group->group_id);
-out_release_minor_instance:
-    ida_simple_remove(&bafs_minor_ida, group->minor);
-out_delete_group_ctrls:
-    for (j = j - 1; j >= 0; j--) {
-        bafs_put_ctrl(group->ctrls[j]);
-    }
-    kfree(group->ctrls);
-out_delete_group:
-    kfree(group);
-out_clean_ctrl_devices:
-    for (i = i - 1; i >= 0; i--) {
-        put_device(ctrl_devices[i]);
-    }
-    put_device(bafs_core_device);
-    kfree (ctrl_devices);
-out_free_local_ctrls:
+    return 0;
+out_free_group:
+    bafs_put_group(group);
+out_free_params:
     kfree(params.ctrls);
 out:
     return ret;
@@ -365,7 +163,7 @@ long bafs_core_delete_group(void __user* user_params) {
     }
 
     put_device(device);
-    bafs_put_group(group, __bafs_group_release);
+    bafs_put_group(group);
 
 out:
     return ret;
@@ -566,20 +364,15 @@ static int __init bafs_init(void) {
     }
 
     //create ctrl class
-    bafs_ctrl_class = class_create(THIS_MODULE, BAFS_CTRL_CLASS_NAME);
-    if (IS_ERR(bafs_ctrl_class)) {
-        ret         = PTR_ERR(bafs_ctrl_class);
-        BAFS_CORE_ERR("Failed to create ctrl class \t err = %d\n", ret);
+    ret = bafs_ctrl_init();
+    if(ret < 0) {
         goto out_destroy_core_class;
     }
 
     //create group class
-    bafs_group_class = class_create(THIS_MODULE, BAFS_GROUP_CLASS_NAME);
-
-    if (IS_ERR(bafs_group_class)) {
-        ret                                                = PTR_ERR(bafs_group_class);
-        BAFS_CORE_ERR("Failed to create group class \t err = %d\n", ret);
-        goto out_destroy_ctrl_class;
+    ret = bafs_group_init();
+    if(ret < 0) {
+        goto out_ctrl_fini;
     }
 
     //init dev objects
@@ -589,7 +382,7 @@ static int __init bafs_init(void) {
     ret = cdev_add(&bafs_core_cdev, MKDEV(MAJOR(bafs_major), BAFS_CORE_MINOR), 1);
     if (ret < 0) {
         BAFS_CORE_ERR("Failed to init core cdev \t err = %d\n", ret);
-        goto out_destroy_group_class;
+        goto out_group_fini;
     }
 
     //create dev
@@ -616,10 +409,12 @@ out_destroy_device:
     device_destroy(bafs_core_class, MKDEV(MAJOR(bafs_major), BAFS_CORE_MINOR));
 out_delete_core_cdev:
     cdev_del(&bafs_core_cdev);
-out_destroy_group_class:
-    class_destroy(bafs_group_class);
-out_destroy_ctrl_class:
-    class_destroy(bafs_ctrl_class);
+out_group_fini:
+    bafs_group_fini();
+
+out_ctrl_fini:
+    bafs_ctrl_fini();
+
 out_destroy_core_class:
     class_destroy(bafs_core_class);
 out_unregister_core_major_region:
@@ -638,8 +433,8 @@ static void __exit bafs_exit(void) {
 
     device_destroy(bafs_core_class, MKDEV(MAJOR(bafs_major), BAFS_CORE_MINOR));
     cdev_del(&bafs_core_cdev);
-    class_destroy(bafs_group_class);
-    class_destroy(bafs_ctrl_class);
+    bafs_group_fini();
+    bafs_ctrl_fini();
     class_destroy(bafs_core_class);
     unregister_chrdev_region(bafs_major, BAFS_MINORS);
