+++ module/bafs/mem.c
@@ -10,8 +10,6 @@
 
 #include <linux/bafs/util.h>
 #include <linux/bafs/types.h>
-#include <linux/bafs/release.h>
-
 
 
 static
